% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMM.R
\name{DTMM_centroid}
\alias{DTMM_centroid}
\title{Getting the cluster centroids}
\usage{
DTMM_centroid(Y, tree, tau_vec = 10^seq(-1, 4, 0.5), nu_vec = 1,
  theta_vec = seq(0.01, 0.99, 0.08), c, gamma)
}
\arguments{
\item{Y}{a \eqn{n * K} matrix containing the OTU/ASV counts of each sample. Each row represents a sample, each column represents an OTU/ASV.}

\item{tree}{an object of class \code{"phylo"} in the package \code{"ape"}. A phylogenetic tree of the OTUs/ASVs.}

\item{tau_vec}{a vector containing the grid of values to numerically integrate out \eqn{\tau}, the cluster-specific dispersion parameter at each node. The values are ordered.}

\item{nu_vec}{a vector containing the grid of values to numerically integrate out \eqn{\nu}, the base dispersion parameter at each node. The values are ordered.}

\item{theta_vec}{a vector containing the grid of values to numerically integrate out \eqn{\theta}, the mean parameter at each node. The values are ordered and are between 0 and 1.}

\item{c}{a vector specifying the clustering labels of the samples that used to estimate the clustering centroids.}

\item{gamma}{a vector specifying the node selection indicators. gamma and c should be set to values in a specific iteration of the Gibbs sampler when fitting DTMM.}
}
\value{
\code{DTMM_centroid} returns a list of length \eqn{K} containing the cluster centroids, where \eqn{K} is the number of distinct values in \eqn{c}.
}
\description{
Get the cluster centroids after fitting DTMM
}
